backend/requirements.txt
# --- Core Frameworks ---
fastapi==0.122.0
uvicorn==0.25.0
pydantic==2.5.2

# --- Database & ODM ---
beanie==2.0.1
motor  # <--- THÊM DÒNG NÀY (Chỉ ghi tên, không ghi version)

# --- Utilities ---
python-dotenv==1.0.0
email-validator>=2.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# --- Caching ---
redis>=5.0.0

# --- Form Data ---
python-multipart==0.0.20
